requirements.txt
# -i https://pypi.org/simple
# numpy==2.3.2; python_version >= '3.11'
# scipy==1.16.1; python_version >= '3.11'
# pyarrow==21.0.0; python_version >= '3.9'

python-dotenv==1.1.1; python_version >= '3.9'
pydantic==2.11.7; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'

psycopg2-binary==2.9.10; python_version >= '3.8'
pandas==2.3.2; python_version >= '3.9'

scikit-learn==1.7.1; python_version >= '3.10'
fastembed==0.7.3; python_full_version >= '3.9.0'
huggingface-hub==0.34.4; python_full_version >= '3.8.0'
openai==1.106.1; python_version >= '3.8'

streamlit==1.49.1; python_version >= '3.9' and python_full_version != '3.9.7'
flask==3.1.2; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
